coupons crud for users

+++ controllers/recycleBinsController.js
@@ -1,5 +1,7 @@
 const RecycleBin = require('../models/recycleBins');
 const { infoLogger, errorLogger } = require("../logs/logs");
+const { userService } = require('../services/userService');
+
 exports.recycleBinsController = {
     getRecycleBins(req, res) {
         infoLogger.info("Get all RecycleBins");
@@ -30,8 +32,17 @@ exports.recycleBinsController = {
                 res.status(500).json({ "message": `Error getting recycle bin` });
             });
     },
-    editRecycleBinDetails(req, res) {
+    async editRecycleBinDetails(req, res) {
         infoLogger.info("Updating a recycleBin");
+        let isMod = false;
+        if (req.userId) {
+            isMod = await userService.isModerator(req.userId);
+        }
+        if (!isMod) {
+            errorLogger.error(`unauthorized user ${req.userId}`);
+            res.status(401).json({ "message": "Unauthorized user" });
+            return;
+        }
         RecycleBin.updateOne({ _id: req.params.id }, req.body)
             .then((result) => {
                 if (result.matchedCount > 0) {
@@ -45,8 +56,17 @@ exports.recycleBinsController = {
             })
             .catch((err) => res.status(400).json({ "message": "Wrong RecycleBin id please enter correct id" }));
     },
-    addRecycleBin(req, res) {
+    async addRecycleBin(req, res) {
         infoLogger.info("Add a recycleBin");
+        let isMod = false;
+        if (req.userId) {
+            isMod = await userService.isModerator(req.userId);
+        }
+        if (!isMod) {
+            errorLogger.error(`unauthorized user ${req.userId}`);
+            res.status(401).json({ "message": "Unauthorized user" });
+            return;
+        }
         const newRecycleBin = new RecycleBin(req.body);
         newRecycleBin.save()
             .then(result => {
@@ -58,8 +78,17 @@ exports.recycleBinsController = {
                 res.status(400).json({ "message": `Error Adding RecycleBin ` });
             });
     },
-    deleteRecycleBin(req, res) {
+    async deleteRecycleBin(req, res) {
         infoLogger.info("Delete a RecycleBin");
+        let isMod = false;
+        if (req.userId) {
+            isMod = await userService.isModerator(req.userId);
+        }
+        if (!isMod) {
+            errorLogger.error(`unauthorized user ${req.userId}`);
+            res.status(401).json({ "message": "Unauthorized user" });
+            return;
+        }
         RecycleBin.deleteOne({ _id: req.params.id })
             .then((result) => {
                 if (result.deletedCount > 0) {
